% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{coef.mvnLinear}
\alias{coef.mvnLinear}
\title{Retrieve Parameter Estimates from an mvnLinear Fit}
\usage{
\method{coef}{mvnLinear}(object, type = NULL)
}
\arguments{
\item{object}{a \code{\link{mvnLinear}} object.}

\item{type}{the type of estimates to return, should be
set to to either "mle" or "naive". If left \code{NULL} then
the mle will be returned.}
}
\value{
An estimate of the mean parameter of the normal
distribution.
}
\description{
Retrieve parameter estimates from an
\code{\link{mvnLinear}} fit.
}
