% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{coef.psatGLM}
\alias{coef.psatGLM}
\title{Retrieve Coefficient Estimates from a psatGLM Fit}
\usage{
\method{coef}{psatGLM}(object, type = NULL, ...)
}
\arguments{
\item{object}{a \code{\link{psatGLM}} object.}

\item{type}{the type of estimates to return, should be
set to to either "mle" or "naive". If left \code{NULL} then
the "mle" will be returned.}
}
\value{
An estimate of the regression coefficients of the
generalized linear model.
}
\description{
Retrieve coefficients estimates from a
\code{\link{psatGLM}} fit.
}
