% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization_functions.R
\name{findUMPU}
\alias{findUMPU}
\title{#' Inference for Normal Means after Aggregate Testing}
\usage{
findUMPU(
  test.stat = NA,
  alpha = NA,
  lower.bound,
  upper.bound,
  mean,
  sd,
  optim.type = "rejection",
  sd.dist = 10,
  optim.control
)
}
\arguments{
\item{test.stat}{the test statistic. Not required for finding the rejection area.}

\item{alpha}{type I error to be kept by the rejection area. Not required for finding p-value.}

\item{lower.bound}{lower bound of the truncated normal distribution.}

\item{upper.bound}{upper bound of the truncated normal distribution.}

\item{mean}{the expected value of the truncated normal distribution.}

\item{sd}{the standard deviation of the truncated normal distribution}

\item{optim.type}{accepts, either rejection or pvalue, indicates which value to find.}

\item{sd.dist}{defines the bounds of the search, will not search for rejection area more than sd.dist Z-scores from the mean.}

\item{optim.control}{allows the user to set some characteristics of the Differential Evolution optimization algorithm implemented in DEoptim.
For more details see code{\link[DEoptim]{DEoptim.control}}.}
}
\description{
\code{findUMPU} is to find either the UMPU rejection area, or the p-value
corresponding to that area.
}
\details{
The function finds either the rejection area or p-value under the conditions of an UMPU test,
as described in section 3.1 \href{https://rss.onlinelibrary.wiley.com/doi/full/10.1111/rssb.12318}{Post-selection estimation and testing following aggregate association tests}.
}
