% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{getCI}
\alias{getCI}
\title{Retrieve Selection Adjusted Confidence Intervals}
\usage{
getCI(object, type = NULL, confidence_level = NULL, switchTune = NULL)
}
\arguments{
\item{object}{an object of class \code{\link{mvnQuadratic}} or
\code{\link{glmQuadratic}}.}

\item{type}{the type of confidence interval to contruct, see the
the documentation of \code{\link{mvnQuadratic}} for details.}

\item{confidence_level}{the desired confidence level for the confidence
intervals}

\item{switchTune}{tuning parameter for Regime Switching confidence intervals.}
}
\value{
a set of confidence intervals for the normal means/regression
coefficients.
}
\description{
A function for retrieving selection adjusted confidence intervals
obtained after aggregate testing using the \code{\link{mvnQuadratic}}
and \code{\link{glmQuadratic}} functions.
}
\details{
This function retrieves confidence intervals from a post-aggregate testing analysis.
if \code{type} is left as \code{NULL} then the confidence interval is the first listed in the
\code{ci_type} field of the original function call. If \code{switchTune} is left
as \code{NULL} then the method used in the original function call will be used.
}
