% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{getPval}
\alias{getPval}
\title{Retrieve Selection Adjusted P-Values}
\usage{
getPval(object, type = NULL)
}
\arguments{
\item{object}{an object of class \code{\link{mvnQuadratic}} or
\code{\link{glmQuadratic}}.}

\item{type}{the type of p-value to compute, see the
the documentation of \code{\link{mvnQuadratic}} for details. If left
\code{NULL} then the first listed value in the \code{pvalue_type} field in
the original function call will be returned.}
}
\value{
a vector of p-values for the normal means/regression
coefficients.
}
\description{
A function for retrieving selection adjusted p-values
obtained after aggregate testing using the \code{\link{mvnQuadratic}}
and \code{\link{glmQuadratic}} functions.
}
