% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Threshold_functions.R
\name{getTestMatEigen}
\alias{getTestMatEigen}
\title{Find Delta of test matrix}
\usage{
getTestMatEigen(cov.mat, K)
}
\arguments{
\item{cov.mat}{The covariance matrix of \code{y}.}

\item{K}{The test matrix \eqn{K} used in the aggregate test}
}
\description{
Function that finds the delta of the test matrix
}
\details{
Finding the weighted chi-square eigen values based on the following
\eqn{y ~ N(\mu, \Sigma)}, and we have the following quadratic form
\eqn{y' K y = y' \Sigma^{-1/2} \Sigma^{1/2} K \Sigma^{1/2} \Sigma^{-1/2} y}.
The functions find the eigen-values of \eqn{\Sigma^{1/2} K \Sigma^{1/2}, where $\Sigma^{1/2}}
is found using cholesky decomposition.
}
