% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdFunctions.R
\name{getTestThreshold}
\alias{getTestThreshold}
\title{Get the Threshold for a General Quadratic Test}
\usage{
getTestThreshold(alpha, sigma, testMat)
}
\arguments{
\item{alpha}{the upper quantile of the distribution/significance
level of the quadratic test.}

\item{sigma}{the covariance of \eqn{y}.}

\item{testMat}{the test matrix \eqn{K}.}
}
\value{
a numeric threshold for the quadratic test.
}
\description{
Computes the upper quantile of the distribution of a quadratic
form: \deqn{y' K y,} where \eqn{y ~ N(0, \Sigma)} and \eqn{K} is
positive semi-definite. This quantile is approximated using the
\code{\link[CompQuadForm]{liu}} method.
}
\seealso{
\code{\link[CompQuadForm]{CompQuadForm}}, \code{\link[CompQuadForm]{liu}}
}
