% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Threshold_functions.R
\name{getThresholdQuadratic}
\alias{getThresholdQuadratic}
\title{Find quantile for weighted Chi-square distribution}
\usage{
getThresholdQuadratic(prob, cov.mat, K, jump.size = 2)
}
\arguments{
\item{prob}{probability}

\item{cov.mat}{The covariance matrix of \code{y}.}

\item{K}{The test matrix \eqn{K} used in the aggregate test. Defaults to \code{solve(cov.mat)}, WALD test.}

\item{jump.size}{parameter for searching the quantile (how fast the increase the searchs)}
}
\description{
\code{getThresholdQuadratic} is used to find a quantile specified of the following test statsitic distribution
\deqn{y' K y \; y \sim N(y, \Sigma)}
}
