% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmInference.R
\name{glmQuadratic}
\alias{glmQuadratic}
\title{Inference for Generalized Linear Models Selected via an Aggregate Test}
\usage{
glmQuadratic(X, y, testMat = "wald", family = "gaussian",
  resid_sd = c("null", "naive"), threshold = NULL, pval_threshold = 0.05,
  estimate_type = c("mle", "naive"), pvalue_type = c("hybrid", "polyhedral",
  "naive", "global-null"), ci_type = c("switch", "polyhedral", "naive",
  "global-null"), confidence_level = 0.95, switchTune = c("sqrd", "half"),
  nSamples = NULL, verbose = TRUE)
}
\arguments{
\item{X}{the design matrix of the regression model without an
intercept.}

\item{y}{the observed dependent variable.}

\item{testMat}{see \code{\link{mvnQuadratic}} for details.}

\item{family}{see \code{\link[stats]{glm}} for details.}

\item{resid_sd}{method for estimating the covariance matrix.}

\item{threshold}{see \code{\link{mvnQuadratic}} for details.}

\item{estimate_type}{see \code{\link{mvnQuadratic}} for details.}

\item{pvalue_type}{see \code{\link{mvnQuadratic}} for details.}

\item{ci_type}{see \code{\link{mvnQuadratic}} for details.}

\item{confidence_level}{see \code{\link{mvnQuadratic}} for details.}

\item{nSamples}{see \code{\link{mvnQuadratic}} for details.}

\item{verbose}{see \code{\link{mvnQuadratic}} for details.}

\item{pthreshold}{see \code{\link{mvnQuadratic}} for details.}

\item{switch_tune}{see \code{\link{mvnQuadratic}} for details.}
}
\value{
an object of type \code{glmQuadrtic}
}
\description{
\code{glmQuadratic} is used to estimate regression models that were selected
based on a single aggregate test of the of form \deqn{\beta' K \beta > c > 0}
where
\eqn{\hat\beta} is the MLE estimate for the regression coefficients.
}
\details{
\code{glmQuadratic} is used to perform inference in generalized
linear models that were selected via a single aggregate test. This
function essentially works as a wrapper for \code{\link{mvnQuadratic}}. See the
documentaiton of \code{\link{mvnQuadratic}} for details regrading the inference
methods.

The function works by fitting a regression model using the \code{\link[stats]{glm}}
function. The covariance matrix is estimated based on the estimated model if
\code{resid_sd} is set to "naive" or based on an intercept only model if
\code{resid_sd} is set to "null". Once a model has been estimated, a quadratic test
is performed based on the estimated covariance with the intercept removed
and the vector of regression coefficients is passed to \code{\link{mvnQuadratic}}
without the intercept.
}
\seealso{
\code{\link{mvnQuadratic}}, \code{\link{getCI}},
\code{\link{getPval}}, \code{\link{plot.glmQuadratic}},
\code{\link{summary.glmQuadratic}}, \code{\link{coef.glmQuadratic}},
\code{\link{predict.glmQuadratic}}
}
