% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalLinearInference.R
\name{mvnLinear}
\alias{mvnLinear}
\title{Inference for Normal Means after Aggregate Testing with a Linear Test}
\usage{
mvnLinear(y, sigma, testVec, threshold = NULL, pval_threshold = 0.05,
  contrasts = NULL, test_direction = c("two-sided", "lower", "upper"),
  estimate_type = c("mle", "naive"), pvalue_type = c("hybrid", "polyhedral",
  "naive"), ci_type = c("switch", "polyhedral", "naive"),
  confidence_level = 0.95, verbose = TRUE, control = psatControl())
}
\arguments{
\item{y}{the observed normal vector.}

\item{sigma}{the covariance matrix of \code{y}.}

\item{testVec}{the test testVec \eqn{a} of size \code{length(y)} used in the aggregate test}

\item{threshold}{the threshold used in the aggregate test, either a vector of size two for the lower and upper
thresholds \eqn{u, l}, or a single number.}

\item{pval_threshold}{the signficance level of the aggregate test.
Overrided by \code{threshold} if both are provided.}

\item{contrasts}{an optional matrix of contrasts to be tested: must have number of columns 
identical to the length of \code{y}. If left as \code{NULL}, the coorinates of \code{y} will be tested by default.}

\item{test_direction}{whether the linear test is one-sided or two-sided. Will be used if the provided threshold 
is a scalar, if lower then the tests will be \eqn{a'y < threshold} and if upper then the test will be \eqn{a'y > threshold}.}

\item{estimate_type}{see \code{\link{mvnQuadratic}} for details.}

\item{pvalue_type}{see \code{\link{mvnQuadratic}} for details.}

\item{ci_type}{see \code{\link{mvnQuadratic}} for details.}

\item{confidence_level}{the confidence level for constructing confidencei intervals.}

\item{verbose}{whether to report on the progress of the computation.}

\item{control}{an object of type \code{\link{psatControl}}.}
}
\value{
An object of class \code{mvnLinear}.
}
\description{
\code{mvnLinear} is used to estimate a normal means model that was selected based
on a single linear aggregate test of the form:
\deqn{a'y > u or a'y < l,} 
\deqn{l < u.}
}
\details{
The function is used to perform inference for normal mean vectors
that were selected based on a single linear aggregate test. To be exact, suppose
that \eqn{y ~ N(\mu,\Sigma)} and that we are interested in estimating \eqn{\mu}
only if we can determine that \eqn{\mu != 0} using an aggregate test of the form:
\eqn{a'y <l} or \eqn{a'y > u} for some predetermined constants \eqn{a, l, u}.

The \code{threshold} parameter specifies the constants \eqn{l<u} which are used
to threshold the aggregate test. If only a single number is provided, then the threshold
will be set according to test_direction: 
\itemize{
\item lower: a'y < threshold
\item upper: a'y > threshold
\item two-sided a'y < -threshold, or a'y > threshold
}
The \code{threshold} parameter takes precedence over \code{pval_threshold} if both
are specified. 

See \code{\link{mvnQuadratic}} for details regarding the available inference methods.
}
\seealso{
\code{\link{mvnQuadratic}}, \code{\link{psatGLM}}, \code{\link{getCI}},
\code{\link{getPval}}.
}
