% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalInference.R
\name{mvnQuadratic}
\alias{mvnQuadratic}
\title{Inference for Normal Means after Aggregate Testing}
\usage{
mvnQuadratic(y, sigma, testMat = "wald", threshold = NULL,
  pval_threshold = 0.05, estimate_type = c("mle", "naive"),
  pvalue_type = c("hybrid", "polyhedral", "naive", "global-null"),
  ci_type = c("switch", "polyhedral", "naive", "global-null"),
  confidence_level = 0.95, switchTune = c("sqrd", "half"),
  nSamples = NULL, verbose = TRUE)
}
\arguments{
\item{y}{the observed normal vector.}

\item{sigma}{the covariance matrix of \code{y}.}

\item{testMat}{the test matrix \eqn{K} used in the aggregate test}

\item{threshold}{the threshold \eqn{c > 0} used in the aggregate test.}

\item{pval_threshold}{the signficance level of the aggregate test.
Overrided by \code{threshold} if both are provided.}

\item{estimate_type}{the types of point estimates to compute and report. The first
estimator listed will be used as the default method.}

\item{pvalue_type}{a vector of methods with which to compute the p-values. The first
method listed will be used as the default method.}

\item{ci_type}{a vector of confidence interval computation methods to be used.
The first method listed will be will be used as the default method.}

\item{confidence_level}{the confidence level for constructing confidencei intervals.}

\item{switchTune}{tuning method for computing the regime switching confidence
intervals}

\item{nSamples}{number of samples to take from the truncated global null distribution.
This only applies if the \code{global-null}, \code{switch} or \code{hybrid} options
are specified.}

\item{verbose}{whether to report on the progress of the computation.}
}
\value{
An object of class \code{mvnQuadratic}.
}
\description{
\code{mvnQuadratic} is used to estimate a normal means model that was selected based
on a single quadratic aggregate test of the form:
\deqn{y' K y > c > 0}
}
\details{
The function is used to perform inference for normal mean vectors
that were selected based on a single quadratic aggregate test. To be exact, suppose
that \eqn{y ~ N(\mu,\Sigma)} and that we are interested in estimating \eqn{\mu}
only if we can determine that \eqn{\mu\neq 0} using an aggregate test of the form:
\deqn{y' K y > c > 0} for some predetermined constant \eqn{c}. If \code{testMat} is set
to the default value of "wald", then \eqn{K = \Sigma^{-1}}. If wald test is used, it is
recommended to specify \code{testMat} as "wald" because this setting makes some of computations
more efficient. Otherwise, \code{testMat} must be a positive definite matrix of an
appropriate dimension.

If \code{estimate_type} includes the string "mle" then \code{mvnQuadratic}
will compute the conditional maximum likelihood estimator for the mean vector,
which is typically a shrinkage estimator.  If \code{testMat = "wald"} then the
computation is performed via an efficient line-search method. Otherwise, a
stochastic gradient method is used.

The \code{threshold} parameter specifies the constant \eqn{c>0} which is used
to threshold the aggregate test. It takes precedence over \code{pval_threshold} if both
are specified. We use the \code{\link[CompQuadForm]{liu}} function to compute the the
threshold if a non-Wald test is used.

\code{mvnQuadratic} offers several options for computing p-values. The "global-null"
method relies on comparing the magnitude of \eqn{y} to samples from the truncated
global-null distribution. This method is powerful when \eqn{\mu} is sparse and its
non-zero coordinates are not very large. The "polyhedral" method is exact when the
observed data is approximately normal and is quite robust to model misspecification.
It tends to be more powerful than the `global-null` method when the magnitude of
\eqn{\mu} is large. The "hybrid" method combines the strengths of the "global-null"
and "polyhedral" methods, possessing good power regardless of the sparsity or
magnitude of \eqn{\mu}. However it is less robust to the misspecification of the distribution
of \eqn{y} than the "polyhedral" method. The confidence interval methods are similar to the p-values ones,
with the Regime switching
confidence intervals ("switch") serving a simialr purpose as the "hybrid" method.

The \code{switchTune} field is used to specifiy how the Regime Switching confidence
intervals should be used. Let \eqn{t1} be the quantile of the aggregate test used for
screening (\code{pval_threshold}) and let \eqn{\alpha} equal one minus \code{confidence_level}.
The regime switching confidence intervals work by performing a secondary aggregate test at
at a quantile \eqn{\gamma} and if the second test is rejected then unadjusted confidence
intervals at a level \eqn{1 - \alpha} are reported. If the second test is not rejected
then global-null confidence intervals at a level \eqn{1 - \alpha + \gamma} are reported.
If \code{switchTune} is set to "sqrd" then \eqn{\gamma = t1 \alpha^{2}} and
\eqn{\gamma = t1 \alpha/2} if \code{switchTune} is set to "half".
}
\seealso{
\code{\link{getCI}}, \code{\link{getPval}},
\code{\link{coef.mvnQuadratic}}, \code{\link{plot.mvnQuadratic}}.
}
