% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plot.glmQuadratic}
\alias{plot.glmQuadratic}
\title{Plot results for glmQuadratic Objects}
\usage{
\method{plot}{glmQuadratic}(object, type = c("estimates", "solution_path",
  "p-values"), ...)
}
\arguments{
\item{object}{and object of class \code{mvnQuadratic} or \code{glmQuadratic}.}

\item{type}{the type of plot to be generated.}

\item{...}{additional arguments to be passed to plotting function.}
}
\value{
a \code{\link[ggplot2]{ggplot2}} figure.
}
\description{
An interface for plotting the results of a post-selection analysis
following an aggregate test. At the moment, methods for plotting
estimates and confidence intervals, p-values and stochastic gradietn solution
paths are implemented.
}
\details{
This functions serves as an interface for plotting the results
of post aggregatet testing inference. The function calls the \code{\link{plotEstimates}}
function for \code{type} "estimates", the \code{\link{plotSolutionPath}} method for
\code{type} "solutionPath" and \code{\link{plotPvalues}} for \code{type} "p-values".
See documentation for these functions for details regarding additional optional parameters.
}
\seealso{
\code{\link{mvnQuadratic}}, \code{\link{glmQuadratic}},
\code{\link{plotEstimates}}, \code{\link{plotSolutionPath}},
 \code{\link{plotPvalues}}
}
