% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plot.mvnLinear}
\alias{plot.mvnLinear}
\title{Plot results of mvnLinear Analyisis}
\usage{
\method{plot}{mvnLinear}(x, type = c("estimates", "solution-path",
  "p-values"), ...)
}
\description{
see \code{\link{plot.glmQuadratic}} for details.
}
