% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plot.mvnQuadratic}
\alias{plot.mvnQuadratic}
\title{Plot results for mvnQuadratic Objects}
\usage{
\method{plot}{mvnQuadratic}(object, type = c("estimates", "solution-path",
  "p-values"), ...)
}
\description{
see \code{\link{plot.glmQuadratic}} for details.
}
