% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotEstimates}
\alias{plotEstimates}
\alias{plotSolutionPath}
\alias{plotPvalues}
\title{Plotting Functions for Post Aggregate Testing Inference Results}
\usage{
plotEstimates(object, true = NULL, offset = 0.12)

plotSolutionPath(object)

plotPvalues(object, threshold = 0.1, adjust = "bonferroni")
}
\arguments{
\item{true}{the true values of the estimated means or regression coefficients.
Mostly useful for experiments with simulated data where the true parameter values
are known.}

\item{threshold}{a vector of thresholds to be plotted along with the p-values.}

\item{adjust}{method for adjusting the p-values. See \code{\link[stats]{p.adjust}}
for details.}
}
\value{
A \code{\link[ggplot2]{ggplot2}} figure.
}
\description{
Functions for plotting the results of post aggregate testing
analyses.
}
\section{Functions}{
\itemize{
\item \code{plotEstimates}: plots estimates and confidence intervals.

\item \code{plotSolutionPath}: plots the solution path of a stochastic
gradient method.

\item \code{plotPvalues}: plots the selection adjusted p-values.
}}

