% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_functions.R
\name{plotFunc}
\alias{plotFunc}
\title{Plotting results of PSAT procedures}
\usage{
plotFunc(ci.df, est.df, truth = NULL, base.line = NULL)
}
\arguments{
\item{ci.df}{data.frame of confidence intervals chosen to plot}

\item{est.df}{data.frame of estimates chosen to plot, `moment` estimates name is changed to `polyhedral.symmetric`}

\item{base.line}{Draw horiziontal line}
}
\description{
\code{plotFunc} is used for basic plotting of the CI and estimates resulting from the PSAT procedure.
}
\details{
Offers some basic plotting of the
}
\examples{
\dontrun{

y         <- c(15, rnorm(p - 1))
cov.mat <- diag(p)
psat.res <- PSATQuadratic(y, cov.mat) ### Runs on default values
plotFunc(psat.res$CI \%>\%

}

}
