% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization_functions.R
\name{polyCIOneSideToOptim}
\alias{polyCIOneSideToOptim}
\title{Returns the difference between the quantile of a truncated normal with a certain probability
to the test statistic.}
\usage{
polyCIOneSideToOptim(mu, stat, lower.trunc, upper.trunc, sd, prob)
}
\description{
Returns the difference between the quantile of a truncated normal with a certain probability
to the test statistic.
}
