% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization_functions.R
\name{polyCIUMPU}
\alias{polyCIUMPU}
\title{#' Inference for Normal Means after Aggregate Testing}
\usage{
polyCIUMPU(stat, lower.trunc, upper.trunc, sd, alpha, sd.dist)
}
\arguments{
\item{stat}{the test statistic.}

\item{lower.trunc}{lower bound of the truncated normal distribution.}

\item{upper.trunc}{upper bound of the truncated normal distribution.}

\item{sd}{the standard deviation of the truncated normal distribution.}

\item{alpha}{\eqn{1 - \alpha} confidence level for the CI.}

\item{sd.dist}{defines the bounds of the search, will not search for rejection area more than sd.dist Z-scores from the mean.
For more details see code{\link[DEoptim]{DEoptim.control}}.}
}
\description{
\code{findUMPU} is to find either the UMPU rejection area, or the p-value
corresponding to that area.
}
\details{
The functions finds the UMAU confidecne interval.
Based on Pratt's theorem \href{https://www.tandfonline.com/doi/abs/10.1080/01621459.1961.10480644}{Pratt Theorem}, instead of directly inverting the UMPU test,
finding the shortest CI will yield the same CI.
}
