% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{predict.glmQuadratic}
\alias{predict.glmQuadratic}
\title{Retrieve Linear Predictors from a glmQuadratic Fit}
\usage{
\method{predict}{glmQuadratic}(object, newX = NULL)
}
\arguments{
\item{object}{a \code{\link{glmQuadratic}} object.}

\item{newX}{a new matrix of covariates. If left \code{NULL} the
linear predictors for the data used for fitting the model will be
returned.}
}
\value{
A vector of linear predictors.
}
\description{
Retrieve the linear predictor from a glmQuadratic
fit. This is the same as the \code{type = "link"} option in
\code{\link[stats]{predict.glm}}.
}
\seealso{
\code{\link{glmQuadratic}}
}
