% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmInference.R
\name{psatGLM}
\alias{psatGLM}
\title{Inference for Generalized Linear Models Selected via an Aggregate Test}
\usage{
psatGLM(X, y, test = "wald", test_direction = c("two-sided", "lower",
  "upper"), family = "gaussian", resid_sd = c("null", "naive"),
  threshold = NULL, pval_threshold = 0.05, estimate_type = c("mle",
  "naive"), pvalue_type = c("hybrid", "polyhedral", "naive"),
  ci_type = c("switch", "polyhedral", "naive"), confidence_level = 0.95,
  verbose = TRUE, control = psatControl())
}
\arguments{
\item{X}{the design matrix of the regression model without an
intercept.}

\item{y}{the observed dependent variable.}

\item{test}{For an aggregate test, the string `wald` or a semi-positive matrix
of dimension \code{ncol(X) X ncol{X}}.
For a linear test a vector of length \code{ncol(X)}.}

\item{test_direction}{If a linear test is specified and no threshold is specififed
then this will be used to determine the threshold.}

\item{family}{see \code{\link[stats]{glm}} for details.}

\item{resid_sd}{method for estimating the covariance matrix. Re-fitted least squares
if `naive` or intercept model if `null`.}

\item{threshold}{a threshold that the aggregate test must cross in order for the model
to be estimated. Should be a positive scalar for quadratic tests and a vector of size 2
for linear aggregate tests.}

\item{pval_threshold}{if threshold is not specified, then a threshold will be set such that
the selection occures iff \code{aggregate_pval < pval_threshold}.}

\item{estimate_type}{see \code{\link{mvnQuadratic}} for details.}

\item{pvalue_type}{see \code{\link{mvnQuadratic}} for details.}

\item{ci_type}{see \code{\link{mvnQuadratic}} for details.}

\item{confidence_level}{see \code{\link{mvnQuadratic}} for details.}

\item{verbose}{whether the progress of the inference function should be reported.}

\item{control}{an object of type \code{\link{psatControl}}.}

\item{switch_tune}{see \code{\link{mvnQuadratic}} for details.}

\item{nSamples}{see \code{\link{mvnQuadratic}} for details.}

\item{verbose}{see \code{\link{mvnQuadratic}} for details.}
}
\value{
an object of type \code{psatGLM}
}
\description{
\code{psatGLM} is used to estimate regression models that were selected
based on a single aggregate test of the of form \deqn{\beta' K \beta > c > 0}
or \deqn{a'\beta < l or a'\beta > u}
where \eqn{\beta} is the MLE estimate for the regression coefficients.
}
\details{
\code{psatGLM} is used to perform inference in generalized
linear models that were selected via a single aggregate test. This
function essentially works as a wrapper for \code{\link{mvnQuadratic}}
and \code{\link{mvnLinear}}. See the
documentation of \code{\link{mvnQuadratic}} for details regrading the
available inference methods.

The function works by fitting a regression model using the \code{\link[stats]{glm}}
function. If \code{resid_sd} is set to "naive" then the covariance matrix is estimated
based on the initial model fit. If \code{resid_sd} is set to "null" then the covariance matrix
will be estimated based on an intercept model. Once a model has been estimated, an
aggregate test is performed based on the estimated covariance with the intercept removed
and the vector of regression coefficients is passed to \code{\link{mvnQuadratic}} or \code{\link{mvnLinear}}
without the intercept.
}
\seealso{
\code{\link{mvnQuadratic}}, \code{\link{getCI}},
\code{\link{getPval}}, \code{\link{plot.psatGLM}},
\code{\link{summary.psatGLM}}, \code{\link{coef.pastGLM}},
\code{\link{predict.psatGLM}}
}
