% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadratic_testing.R
\name{PSATQuadratic}
\alias{PSATQuadratic}
\title{Conduct Post Selection After Qudratic Aggregation Testing inference}
\usage{
PSATQuadratic(
  y,
  cov.mat,
  alpha = 0.05,
  K = solve(cov.mat),
  threshold = qchisq(1 - alpha, nrow(cov.mat)),
  contrast = NULL,
  alternative = NULL,
  test.type = c("symmetric", "UMPU"),
  pval.type = c("switch", "global", "hybrid", "polyhedral", "naive"),
  ci.type = pval.type,
  est.type = c("naive", "moment"),
  verbose = FALSE,
  zscore.bound = 5,
  samp.para = list(samp.size = 10000, burn.in = 200),
  optim.control = DEoptim::DEoptim.control()
)
}
\arguments{
\item{y}{The observed normal vector.}

\item{cov.mat}{The covariance matrix of \code{y}.}

\item{alpha}{Indicates the type I error level in which to find rejection areas and confidence intervals.}

\item{K}{The test matrix \eqn{K} used in the aggregate test. Defaults to \code{solve(cov.mat)}, WALD test.}

\item{threshold}{The threshold \eqn{c > 0} used in the aggregate test.}

\item{contrast}{An optional matrix of contrasts to be tested: must have number of columns
identical to the length of \code{y}. If left as \code{NULL}, the coordinates of \code{y} will be tested by default.}

\item{alternative}{A string.vector indicated the for which contrast what type of test to conduct
\code{'two.sided'}, \code{'less'} or \code{'greater'}. Should have the same length as number of contrasts.
Defaults to two.sided tests.}

\item{test.type}{Type of tests (UMPU and Symmetric), only relevant for two-sided testing.}

\item{pval.type}{Type of p-values to calculate for each contrast.
Options available are \code{'global', 'hybrid', 'polyhedral', 'naive'}.}

\item{ci.type}{Type of confidence intervals to create for each contrast.
Options available are \code{'global', 'hybrid', 'polyhedral', 'switch', 'naive'}.}

\item{est.type}{Type of point estimation, options available are \code{'moment', 'naive'}.}

\item{verbose}{Whether to report on the progress of the computation.}

\item{zscore.bound}{Number of the furthest bound before ignoring truncation and addressing the distribution as normal.
Defaults to the \code{pval.type}.}

\item{samp.para}{List of parameters for the tmg truncated normal sampler.
Contains two parameters \code{samp.size} and \code{burn.in}, for more details see \code{tmg::rtmg}.}

\item{optim.control}{allows the user to set some characteristics of the Differential Evolution optimization algorithm implemented in DEoptim.
For more details see code{\link[DEoptim]{DEoptim.control}}.}
}
\description{
\code{PSATQuadratic} is used to estimate (point & interval) and inferon normal means model that was selected based
on a single quadratic aggregate test of the form:
\deqn{y' K y > c > 0}
}
\details{
The function is used to perform inference for normal mean vectors
that were selected based on a single quadratic aggregate test. To be exact, suppose
that \eqn{y ~ N(\mu,\Sigma)} and that we are interested in estimating \eqn{\mu}
only if we can determine that \eqn{\mu\neq 0} using an aggregate test of the form:
\deqn{y' K y > c > 0} for some predetermined constant \eqn{c}.

\code{PSATQuadratic} offers several options for computing p-values. The "global-null"
method relies on comparing the magnitude of \eqn{y} to samples from the truncated
global-null distribution. This method is powerful when \eqn{\mu} is sparse and its
non-zero coordinates are not very large. The `polyhedral` method is exact when the
observed data is approximately normal and is quite robust to model misspecification.
It tends to be more powerful than the `global-null` method when the magnitude of
\eqn{\mu} is large. The "hybrid" method combines the strengths of the "global-null"
and `polyhedral` methods, possessing good power regardless of the sparsity or
magnitude of \eqn{\mu}. However it is less robust to the misspecification of the distribution
of \eqn{y} than the "polyhedral" method. The confidence interval methods are similar to the p-values ones.
For more details see \href{https://rss.onlinelibrary.wiley.com/doi/full/10.1111/rssb.12318}{Post-selection estimation and testing following aggregate association tests}.
}
\examples{
\dontrun{
y         <- c(15, rnorm(p - 1))
cov.mat <- diag(p)
PSATQuadratic(y, cov.mat) ### Runs on default values
}

\dontrun{
y         <- c(10, rnorm(p - 1))
contrast  <- rbind(rep(1, p), c(1, 1, 1, rep(0, p - 3)))
PSATQuadratic(y = y, cov.mat = cov.mat,
              contrast = contrast, alternative =  c('less', 'greater')) ## Test different alternatives
}
}
