% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{summary.glmQuadratic}
\alias{summary.glmQuadratic}
\title{Summarizing glmQuadratic Fits}
\usage{
\method{summary}{glmQuadratic}(object, estimate_type = NULL,
  pvalue_type = NULL, ci_type = NULL, confidence_level = NULL)
}
\arguments{
\item{object}{an object of type \code{\link{glmQuadratic}}.}

\item{estimate_type}{see \code{\link{mvnQuadratic}} for details.}

\item{pvalue_type}{see \code{\link{mvnQuadratic}} for details.}

\item{ci_type}{see \code{\link{mvnQuadratic}} for details.}

\item{confidence_level}{see \code{\link{mvnQuadratic}} for details.}
}
\value{
A list containing a table of regression coefficients and details
regarding the inference methods used.
}
\description{
A summary method for \code{\link{glmQuadratic}} objects.
}
\details{
This is a summary method for summarizing the results of
post aggregate testing analysis with the \code{\link{glmQuadratic}}
method. The main output of the function is a table of regression coefficients
with confidence intervals and p-values computed using the desired methods.
This function is accompanied by a convenient printing function.
}
\seealso{
\code{\link{glmQuadratic}}
}
