% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Independent_testing.R
\name{aggregatePvalues}
\alias{aggregatePvalues}
\title{Post Selection with Aggregate Testing for Independent P-Values}
\usage{
aggregatePvalues(
  p.mat,
  global.test = c("Fisher", "Pearson"),
  pval.threshold = NULL
)
}
\arguments{
\item{p.mat}{An \code{MXn} matrix of original p-values , where \code{M} is
the number of features and \code{n} is the number of independent p-values
for each feature. If \code{global.test="Pearson"}, the original p-values must
be one-sided p-values. No default.}

\item{global.test}{"Fisher" or "Pearson". The default is "Fisher". Set it to
"Pearson" only for two-sided alternatives, if common directionality is expected
for each feature.}

\item{pval.threshold}{The p-value selection threshold.
If \code{pval.threshold=NULL}, only the global null p-value is computed. Must be
entered for computation of the conditional p-values. Either a scalar, or a vector
of length M with positive entries. See Details.}
}
\value{
List containing the global null p-value in \code{pF},  and the
conditional p-values in \code{p2C}. If a global null p-value is above
the threshold, then the conditional p-value vector corresponding to it is
a vector of NA's; if \code{pval.threshold=NULL}, then the conditional p-value matrix
P2C is NA's.
}
\description{
Performs selection adjustment for independent p-values
following selection with an aggregate level test statistic.
}
\details{
The Fisher test statisic is minus two times the sum of the log of
the input p-values, and the Fisher global null p-value is the probability
that a chi-squared distribution with \code{2n} degrees of freedom exceeds
the Fisher test statistc, where \code{n} is the number of input p-values.

The Pearson test statistic is the maximum of the Fisher test statistic based
on left-sided p-values and the Fisher test statistic based on right-sided
p-values, and the Pearson global null p-value is twice the probability that
a chi-squared distribution with \code{2n} degrees of freedom exceeds the
Pearson test statistic. The global null p-value is output in pF. If the
hypotheses are one-sided, set global.test= "Fisher"; if the hypotheses are
two-sided, set global.test="Pearson". It is assumed that the p-values come from
continuous test statistics, so that left sided pvalue +  right sided pvalue = 1.
See Heller et al. (2016) for details about these global tests, about the
post-selection inference using these tests, and about extensions to other
global tests.

The p-value threshold, pval.threshold, is the threshold that the global null
p-value has to reach in order to be selected for post-selection inference.
For example,  for a bonferroni correction at level \code{alpha} for a family
of \code{m} global null hypotheses, it should be set to \code{alpha/m}. The
conditional p-values are computed only if \code{pF<=pval.threshold}.
}
\references{
Ruth Heller, Nilanjan Chatterjee, Abba Krieger, Jianxin Shi
(2016). Post-selection Inference Following Aggregate Level Hypothesis
Testing in Large Scale Genomic Data. bioRxiv:
\url{http://dx.doi.org/10.1101/058404}
}
